requirements.txt
flask
flask-cors
pandas
qdrant-client
python-dotenv
numpy
google-generativeai
